requirements.txt
Django
django-cors-headers
django-environ
django-model-utils
djangorestframework
drf-secure-token
mysqlclient
pillow

